override for lms/index.html template (#6)
Change welcome message to MITx

removed register link from top nav

adding favicon for lms

Added rule to override lowercasing of second line

The second line in the header title is being lowercased. Added a rule to
overrides to prevent that from happening as it looks strange to have it
fully downcased.